Reorganize for Visual Studio.
Start adding socket support.

+++ json.hpp
@@ -262,7 +262,7 @@ class JSON
 
         int length() const {
             if( Type == Class::Array )
-                return Internal.List->size();
+                return (int) Internal.List->size();
             else
                 return -1;
         }
@@ -275,9 +275,9 @@ class JSON
 
         int size() const {
             if( Type == Class::Object )
-                return Internal.Map->size();
+                return (int) Internal.Map->size();
             else if( Type == Class::Array )
-                return Internal.List->size();
+                return (int) Internal.List->size();
             else
                 return -1;
         }
